+++ src/xqt-kotlinx-lsp/commonTest/xqt/kotlinx/lsp/test/textDocument/CompletionItem.kt
@@ -33,6 +33,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -56,6 +57,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -79,6 +81,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -102,6 +105,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -125,6 +129,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -148,6 +153,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -171,6 +177,7 @@ class TheCompletionItemType {
         assertEquals("Lorem ipsum dolor", params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
@@ -204,8 +211,9 @@ class TheCompletionItemType {
         assertEquals(null, params.sortText)
         assertEquals(null, params.filterText)
         assertEquals(null, params.insertText)
-        assertEquals(null, params.data)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
+        assertEquals(null, params.data)
 
         assertEquals(Position(5u, 23u), params.textEdit?.range?.start)
         assertEquals(Position(5u, 23u), params.textEdit?.range?.end)
@@ -245,12 +253,43 @@ class TheCompletionItemType {
         assertEquals(null, params.filterText)
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
+        assertEquals(null, params.command)
         assertEquals(null, params.data)
 
         val textEdit = params.additionalTextEdits[0]
         assertEquals(Position(5u, 23u), textEdit.range.start)
         assertEquals(Position(5u, 23u), textEdit.range.end)
         assertEquals("Lorem ipsum dolor", textEdit.newText)
+
+        assertEquals(json, CompletionItem.serializeToJson(params))
+    }
+
+    @Test
+    @DisplayName("supports the command property")
+    fun supports_the_command_property() {
+        val json = jsonObjectOf(
+            "label" to JsonPrimitive("Lorem Ipsum"),
+            "command" to jsonObjectOf(
+                "title" to JsonPrimitive("save"),
+                "command" to JsonPrimitive("file.save")
+            )
+        )
+
+        val params = CompletionItem.deserialize(json)
+        assertEquals("Lorem Ipsum", params.label)
+        assertEquals(null, params.kind)
+        assertEquals(null, params.detail)
+        assertEquals(null, params.documentation)
+        assertEquals(null, params.sortText)
+        assertEquals(null, params.filterText)
+        assertEquals(null, params.insertText)
+        assertEquals(null, params.textEdit)
+        assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.data)
+
+        assertEquals("save", params.command?.title)
+        assertEquals("file.save", params.command?.command)
+        assertEquals(null, params.command?.arguments)
 
         assertEquals(json, CompletionItem.serializeToJson(params))
     }
@@ -273,6 +312,7 @@ class TheCompletionItemType {
         assertEquals(null, params.insertText)
         assertEquals(null, params.textEdit)
         assertEquals(0, params.additionalTextEdits.size)
+        assertEquals(null, params.command)
         assertEquals(JsonPrimitive("Lorem ipsum dolor"), params.data)
 
         assertEquals(json, CompletionItem.serializeToJson(params))